undo `fade_out` prop being a bool as it causes a byte size diff

+++ structs/src/scly_props/actor_keyframe.rs
@@ -14,7 +14,7 @@ pub struct ActorKeyframe<'r> {
     pub loop_: u8,
     pub loop_duration: f32,
     pub active: u8,
-    pub fade_out: u8,
+    pub fade_out: u32,
     pub playback_rate: f32,
 }
